# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "super movies api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/device/{public_id}/rpc": {
      "post": {
        "tags": [
          "remote"
        ],
        "summary": "Send",
        "description": "Press a button on somebody's television.\n\n`202`, not `200`: this hands the command to whoever is listening and does\nnot wait to hear how it went. `listeners` in the answer is the honest\nversion of \"did that work\" — zero means nobody was connected.",
        "operationId": "send_device__public_id__rpc_post",
        "parameters": [
          {
            "name": "public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Public Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/device/{public_id}/events": {
      "get": {
        "tags": [
          "remote"
        ],
        "summary": "Watch",
        "description": "A remote watching what a box is doing.\n\nOpens with whatever the box last said, so a phone that just unlocked has\nsomething to draw before the next change arrives.",
        "operationId": "watch_device__public_id__events_get",
        "parameters": [
          {
            "name": "public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Public Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/device/state": {
      "post": {
        "tags": [
          "remote"
        ],
        "summary": "Report",
        "description": "The box saying what it is doing.",
        "operationId": "report_device_state_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommandIn": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "method": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_.]*$",
            "title": "Method"
          },
          "params": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "id",
          "method"
        ],
        "title": "CommandIn",
        "description": "One button press.\n\n`id` is the remote's own, and comes back in the answer — without it a remote\ncannot tell a command that was lost from one that was ignored, and cannot\nsafely resend."
      },
...
      "DeliveryOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "listeners": {
            "type": "integer",
            "title": "Listeners"
          }
        },
        "type": "object",
        "required": [
          "id",
          "listeners"
        ],
        "title": "DeliveryOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonValue": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "StateIn": {
        "properties": {
          "state": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "State"
          }
        },
        "type": "object",
        "title": "StateIn",
        "description": "What the box is doing. The server does not interpret it."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}